+++ Source/DREAM3DLib/IOFilters/ReadOrientationData.h
@@ -178,7 +178,7 @@ class DREAM3DLib_EXPORT ReadOrientationData : public AbstractFilter
 
       DataArray<unsigned int>::Pointer crystalStructures = DataArray<unsigned int>::CreateArray(phases.size() + 1, m_CrystalStructuresArrayName);
       StringDataArray::Pointer materialNames = StringDataArray::CreateArray(phases.size() + 1, m_MaterialNameArrayName);
-      QVector<int> dims(1, 6);
+      QVector<size_t> dims(1, 6);
       FloatArrayType::Pointer latticeConstants = FloatArrayType::CreateArray(phases.size() + 1, dims, m_LatticeConstantsArrayName);
 
       // Initialize the zero'th element to unknowns. The other elements will
