+++ Source/DREAM3DLib/SurfaceMeshingFilters/TriangleAreaFilter.cpp
@@ -171,7 +171,7 @@ void TriangleAreaFilter::dataCheck()
   }
   else
   {
-    QVector<int> dims(1, 1);
+    QVector<size_t> dims(1, 1);
     m_SurfaceMeshTriangleAreasPtr = faceAttrMat->createNonPrereqArray<DataArray<double>, AbstractFilter, double>(this, m_SurfaceMeshTriangleAreasArrayName, 0, dims); /* Assigns the shared_ptr<> to an instance variable that is a weak_ptr<> */
     if( NULL != m_SurfaceMeshTriangleAreasPtr.lock().get() ) /* Validate the Weak Pointer wraps a non-NULL pointer to a DataArray<T> object */
     { m_SurfaceMeshTriangleAreas = m_SurfaceMeshTriangleAreasPtr.lock()->getPointer(0); } /* Now assign the raw pointer to data from the DataArray<T> object */
