+++ Source/DREAM3DLib/SurfaceMeshingFilters/TriangleDihedralAngleFilter.cpp
@@ -196,7 +196,7 @@ void TriangleDihedralAngleFilter::dataCheck()
   }
   else
   {
-    QVector<int> dims(1, 1);
+    QVector<size_t> dims(1, 1);
     m_SurfaceMeshTriangleDihedralAnglesPtr = faceAttrMat->createNonPrereqArray<DataArray<double>, AbstractFilter, double>(this, m_SurfaceMeshTriangleDihedralAnglesArrayName, 0, dims); /* Assigns the shared_ptr<> to an instance variable that is a weak_ptr<> */
     if( NULL != m_SurfaceMeshTriangleDihedralAnglesPtr.lock().get() ) /* Validate the Weak Pointer wraps a non-NULL pointer to a DataArray<T> object */
     { m_SurfaceMeshTriangleDihedralAngles = m_SurfaceMeshTriangleDihedralAnglesPtr.lock()->getPointer(0); } /* Now assign the raw pointer to data from the DataArray<T> object */
