+++ Source/Plugins/ImageImportPlugin/ImageImportPluginFilters/SaveImages.cpp
@@ -176,7 +176,7 @@ void SaveImages::dataCheck()
   }
   else
   {
-    QVector<int> dims(1, 3);
+    QVector<size_t> dims(1, 3);
     m_ColorsPtr = cellAttrMat->getPrereqArray<DataArray<uint8_t>, AbstractFilter>(this, m_ColorsArrayName, -300, dims); /* Assigns the shared_ptr<> to an instance variable that is a weak_ptr<> */
     if( NULL != m_ColorsPtr.lock().get() ) /* Validate the Weak Pointer wraps a non-NULL pointer to a DataArray<T> object */
     { m_Colors = m_ColorsPtr.lock()->getPointer(0); } /* Now assign the raw pointer to data from the DataArray<T> object */
